refactor components to make them reuable
