+++ application/react/DeliveryMethod/DeliveryMethodTable/ExpandedRowContent/ExpandedRowContent.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import './ExpandedRowContent.scss'
+
+const ExpandedRowContent = () => {
+  return (
+    <div>TODO: ExpandedRowContent.js</div>
+  )
+}
+
+export default ExpandedRowContent
